[ALS-6103] Refactor FENCEAuthenticationService
The FENCEAuthenticationService is no longer responsible for making Roles, Privileges, and AccessRules. The functionality has been moved to related services. Additionally, migrated the word "FENCE" to "MANAGED" as refactor our psama application we want a more generic naming convention. Roles, Privileges, and AccessRules are now either MANUAL or MANAGED.
